requirements.txt
absl-py
appdirs
astor
certifi
chardet
coreapi
coreschema
distlib
Django
django-cors-headers
djangorestframework
filelock
gast
google-pasta
grpcio
h5py
idna
importlib-metadata
importlib-resources
itypes
Jinja2
Keras
Keras-Applications
Keras-Preprocessing
Markdown
MarkupSafe
numpy
opt-einsum
Pillow
pipenv
protobuf
pytz
PyYAML
requests
scipy
six
sqlparse
tensorboard
tensorflow
tensorflow-estimator
termcolor
uritemplate
urllib3
virtualenv
virtualenv-clone
Werkzeug
wrapt
zipp
inference
